requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Google Cloud Services
google-cloud-aiplatform>=1.38.0
google-cloud-dialogflow-cx>=1.20.0
google-cloud-vision>=3.5.0
google-cloud-pubsub>=2.18.0
google-cloud-firestore>=2.13.0
google-cloud-storage>=2.14.0
google-cloud-translate>=3.13.0
google-cloud-speech>=2.23.0

# Web Framework & API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
aiohttp>=3.9.0
httpx>=0.25.0
requests>=2.31.0

# Machine Learning & AI
tensorflow>=2.15.0
scikit-learn>=1.3.2
numpy>=1.26.0
pandas>=2.1.0
transformers>=4.35.0
torch>=2.1.0

# Data Validation & Processing
pydantic>=2.5.0
marshmallow>=3.20.0

# Async & Concurrency
asyncio>=3.4.3
aiofiles>=23.2.1

# Database & Caching
redis>=5.0.1
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9

# Logging & Monitoring
loguru>=0.7.2
python-json-logger>=2.0.7

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
pytest-mock>=3.12.0
responses>=0.24.1
faker>=20.1.0
freezegun>=1.4.0

# Development Tools
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.1
pylint>=3.0.3
pre-commit>=3.6.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
geopy>=2.4.1
pillow>=10.1.0
